Move section components to sections
